+++ src/type/IterableType.php
@@ -61,6 +61,9 @@ final class IterableType extends Type
         return $this->allowsNull;
     }
 
+    /**
+     * @psalm-assert-if-true IterableType $this
+     */
     public function isIterable(): bool
     {
         return true;