+++ src/type/ObjectType.php
@@ -57,6 +57,9 @@ final class ObjectType extends Type
         return $this->className;
     }
 
+    /**
+     * @psalm-assert-if-true ObjectType $this
+     */
     public function isObject(): bool
     {
         return true;