+++ src/type/StaticType.php
@@ -51,6 +51,9 @@ final class StaticType extends Type
         return $this->allowsNull;
     }
 
+    /**
+     * @psalm-assert-if-true StaticType $this
+     */
     public function isStatic(): bool
     {
         return true;