+++ tests/unit/ReflectionMapperTest.php
@@ -20,7 +20,12 @@ use ReflectionMethod;
 use SebastianBergmann\Type\TestFixture\AnInterface;
 use SebastianBergmann\Type\TestFixture\AnotherInterface;
 use SebastianBergmann\Type\TestFixture\ChildClass;
+use SebastianBergmann\Type\TestFixture\ClassWithMethodsThatDeclareDisjunctiveNormalFormParameterTypes;
+use SebastianBergmann\Type\TestFixture\ClassWithMethodsThatDeclareDisjunctiveNormalFormReturnTypes;
+use SebastianBergmann\Type\TestFixture\ClassWithMethodsThatDeclareIntersectionParameterTypes;
+use SebastianBergmann\Type\TestFixture\ClassWithMethodsThatDeclareParameterTypes;
 use SebastianBergmann\Type\TestFixture\ClassWithMethodsThatDeclareReturnTypes;
+use SebastianBergmann\Type\TestFixture\ClassWithMethodsThatDeclareUnionParameterTypes;
 use SebastianBergmann\Type\TestFixture\ClassWithMethodsThatDeclareUnionReturnTypes;
 use SebastianBergmann\Type\TestFixture\ClassWithMethodsThatHaveStaticReturnTypes;
 use SebastianBergmann\Type\TestFixture\ClassWithMethodThatDeclaresFalseReturnType;
@@ -46,10 +51,11 @@ use SebastianBergmann\Type\TestFixture\ParentClass;
 #[UsesClass(UnionType::class)]
 #[UsesClass(UnknownType::class)]
 #[UsesClass(VoidType::class)]
+#[UsesClass(Parameter::class)]
 #[Small]
 final class ReflectionMapperTest extends TestCase
 {
-    #[DataProvider('types')]
+    #[DataProvider('typeProvider')]
     public function testMapsFromReturnType(string $expected, ReflectionFunction|ReflectionMethod $method): void
     {
         $this->assertSame($expected, (new ReflectionMapper)->fromReturnType($method)->name());
@@ -156,7 +162,85 @@ final class ReflectionMapperTest extends TestCase
         $this->assertSame('null', $type->name());
     }
 
-    public function types(): array
+    #[RequiresPhp('>= 8.2')]
+    public function testMapsFromDisjunctiveNormalFormReturnType(): void
+    {
+        $type = (new ReflectionMapper)->fromReturnType(new ReflectionMethod(ClassWithMethodsThatDeclareDisjunctiveNormalFormReturnTypes::class, 'one'));
+        $this->assertInstanceOf(UnionType::class, $type);
+        $this->assertSame('(SebastianBergmann\Type\TestFixture\A&SebastianBergmann\Type\TestFixture\B)|SebastianBergmann\Type\TestFixture\D', $type->name());
+
+        $type = (new ReflectionMapper)->fromReturnType(new ReflectionMethod(ClassWithMethodsThatDeclareDisjunctiveNormalFormReturnTypes::class, 'two'));
+        $this->assertInstanceOf(UnionType::class, $type);
+        $this->assertSame('(SebastianBergmann\Type\TestFixture\D&SebastianBergmann\Type\TestFixture\X)|SebastianBergmann\Type\TestFixture\C|null', $type->name());
+
+        $type = (new ReflectionMapper)->fromReturnType(new ReflectionMethod(ClassWithMethodsThatDeclareDisjunctiveNormalFormReturnTypes::class, 'three'));
+        $this->assertInstanceOf(UnionType::class, $type);
+        $this->assertSame('(SebastianBergmann\Type\TestFixture\A&SebastianBergmann\Type\TestFixture\B&SebastianBergmann\Type\TestFixture\D)|int|null', $type->name());
+    }
+
+    public function testMapsFromParameters(): void
+    {
+        $method = new ReflectionMethod(ClassWithMethodsThatDeclareParameterTypes::class, 'unknown');
+        $types  = (new ReflectionMapper)->fromParameterTypes($method);
+
+        $this->assertCount(1, $types);
+        $this->assertSame('x', $types[0]->name());
+        $this->assertSame('', $types[0]->type()->asString());
+
+        $method = new ReflectionMethod(ClassWithMethodsThatDeclareParameterTypes::class, 'named');
+        $types  = (new ReflectionMapper)->fromParameterTypes($method);
+
+        $this->assertCount(1, $types);
+        $this->assertSame('x', $types[0]->name());
+        $this->assertSame('SebastianBergmann\Type\TestFixture\A', $types[0]->type()->asString());
+    }
+
+    public function testMapsFromUnionTypeParameters(): void
+    {
+        $method = new ReflectionMethod(ClassWithMethodsThatDeclareUnionParameterTypes::class, 'union');
+        $types  = (new ReflectionMapper)->fromParameterTypes($method);
+
+        $this->assertCount(1, $types);
+        $this->assertSame('x', $types[0]->name());
+        $this->assertSame('bool|int', $types[0]->type()->asString());
+    }
+
+    public function testMapsFromIntersectionTypeParameters(): void
+    {
+        $method = new ReflectionMethod(ClassWithMethodsThatDeclareIntersectionParameterTypes::class, 'intersection');
+        $types  = (new ReflectionMapper)->fromParameterTypes($method);
+
+        $this->assertCount(1, $types);
+        $this->assertSame('x', $types[0]->name());
+        $this->assertSame('SebastianBergmann\Type\TestFixture\A&SebastianBergmann\Type\TestFixture\B', $types[0]->type()->asString());
+    }
+
+    #[RequiresPhp('>= 8.2')]
+    public function testMapsFromDisjunctiveNormalFormParameters(): void
+    {
+        $method = new ReflectionMethod(ClassWithMethodsThatDeclareDisjunctiveNormalFormParameterTypes::class, 'dnfOne');
+        $types  = (new ReflectionMapper)->fromParameterTypes($method);
+
+        $this->assertCount(1, $types);
+        $this->assertSame('x', $types[0]->name());
+        $this->assertSame('(SebastianBergmann\Type\TestFixture\A&SebastianBergmann\Type\TestFixture\B)|SebastianBergmann\Type\TestFixture\D', $types[0]->type()->asString());
+
+        $method = new ReflectionMethod(ClassWithMethodsThatDeclareDisjunctiveNormalFormParameterTypes::class, 'dnfTwo');
+        $types  = (new ReflectionMapper)->fromParameterTypes($method);
+
+        $this->assertCount(1, $types);
+        $this->assertSame('x', $types[0]->name());
+        $this->assertSame('(SebastianBergmann\Type\TestFixture\D&SebastianBergmann\Type\TestFixture\X)|SebastianBergmann\Type\TestFixture\C|null', $types[0]->type()->asString());
+
+        $method = new ReflectionMethod(ClassWithMethodsThatDeclareDisjunctiveNormalFormParameterTypes::class, 'dnfThree');
+        $types  = (new ReflectionMapper)->fromParameterTypes($method);
+
+        $this->assertCount(1, $types);
+        $this->assertSame('x', $types[0]->name());
+        $this->assertSame('(SebastianBergmann\Type\TestFixture\A&SebastianBergmann\Type\TestFixture\B&SebastianBergmann\Type\TestFixture\D)|int|null', $types[0]->type()->asString());
+    }
+
+    public function typeProvider(): array
     {
         return [
             [