+++ tests/unit/type/UnionTypeTest.php
@@ -14,32 +14,55 @@ use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Small;
 use PHPUnit\Framework\Attributes\UsesClass;
 use PHPUnit\Framework\TestCase;
+use SebastianBergmann\Type\TestFixture\AnInterface;
+use SebastianBergmann\Type\TestFixture\AnotherInterface;
 
 #[CoversClass(UnionType::class)]
 #[CoversClass(Type::class)]
 #[UsesClass(NullType::class)]
 #[UsesClass(SimpleType::class)]
+#[UsesClass(IntersectionType::class)]
+#[UsesClass(ObjectType::class)]
+#[UsesClass(TypeName::class)]
 #[Small]
 final class UnionTypeTest extends TestCase
 {
-    public function testCanBeRepresentedAsString(): void
+    #[DataProvider('stringRepresentationProvider')]
+    public function testCanBeRepresentedAsString(string $expected, Type $type): void
     {
-        $type = new UnionType(
-            Type::fromName('bool', false),
-            Type::fromName('int', false)
-        );
-
-        $this->assertSame('bool|int', $type->asString());
+        $this->assertSame($expected, $type->asString());
     }
 
-    public function testTypesOfUnionAreSortedByNameInStringRepresentation(): void
+    public function stringRepresentationProvider(): array
     {
-        $type = new UnionType(
-            Type::fromName('int', false),
-            Type::fromName('bool', false)
-        );
+        return [
+            [
+                'bool|int',
+                new UnionType(
+                    Type::fromName('bool', false),
+                    Type::fromName('int', false)
+                ),
+            ],
+            [
+                '(' . AnInterface::class . '&' . AnotherInterface::class . ')|bool',
+                new UnionType(
+                    new IntersectionType(
+                        Type::fromName(AnInterface::class, false),
+                        Type::fromName(AnotherInterface::class, false)
+                    ),
+                    Type::fromName('bool', false)
+                ),
+            ],
+        ];
+    }
 
-        $this->assertSame('bool|int', $type->asString());
+    public function testTypesCanBeQueried(): void
+    {
+        $bool = Type::fromName('bool', false);
+        $null = Type::fromName('null', true);
+        $type = new UnionType($bool, $null);
+
+        $this->assertSame([$bool, $null], $type->types());
     }
 
     public function testMayAllowNull(): void
@@ -62,13 +85,36 @@ final class UnionTypeTest extends TestCase
         $this->assertFalse($type->allowsNull());
     }
 
-    #[DataProvider('assignableTypes')]
+    public function testMayContainIntersectionType(): void
+    {
+        $type = new UnionType(
+            Type::fromName('bool', false),
+            new IntersectionType(
+                Type::fromName(AnInterface::class, false),
+                Type::fromName(AnotherInterface::class, false)
+            ),
+        );
+
+        $this->assertTrue($type->containsIntersectionTypes());
+    }
+
+    public function testMayNotContainIntersectionType(): void
+    {
+        $type = new UnionType(
+            Type::fromName('bool', false),
+            Type::fromName('int', false)
+        );
+
+        $this->assertFalse($type->containsIntersectionTypes());
+    }
+
+    #[DataProvider('assignableProvider')]
     public function testAssignableTypesAreRecognized(bool $expected, Type $type, UnionType $union): void
     {
         $this->assertSame($expected, $union->isAssignable($type));
     }
 
-    public function assignableTypes(): array
+    public function assignableProvider(): array
     {
         return [
             [