+++ ChangeLog.md
@@ -8,7 +8,16 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 * This component is no longer supported on PHP 7.3, PHP 7.4 and PHP 8.0
 
-## [3.1.0] - 2022-MM-DD
+## [3.2.0] - 2022-MM-DD
+
+### Added
+
+* [#25](https://github.com/sebastianbergmann/type/issues/25): Support Disjunctive Normal Form types
+* Added `ReflectionMapper::fromParameterTypes()`
+* Added `IntersectionType::types()` and `UnionType::types()`
+* Added `UnionType::containsIntersectionTypes()`
+
+## [3.1.0] - 2022-08-29
 
 ### Added
 
@@ -136,8 +145,9 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 * Initial release based on [code contributed by Michel Hartmann to PHPUnit](https://github.com/sebastianbergmann/phpunit/pull/3673)
 
-[4.0.0]: https://github.com/sebastianbergmann/type/compare/3.1...main
-[3.1.0]: https://github.com/sebastianbergmann/type/compare/3.0.0...3.1
+[4.0.0]: https://github.com/sebastianbergmann/type/compare/3.2...main
+[3.2.0]: https://github.com/sebastianbergmann/type/compare/3.1.0...3.2
+[3.1.0]: https://github.com/sebastianbergmann/type/compare/3.0.0...3.1.0
 [3.0.0]: https://github.com/sebastianbergmann/type/compare/2.3.4...3.0.0
 [2.3.4]: https://github.com/sebastianbergmann/type/compare/ca39369c41313ed12c071ed38ecda8fcdb248859...2.3.4
 [2.3.3]: https://github.com/sebastianbergmann/type/compare/2.3.2...ca39369c41313ed12c071ed38ecda8fcdb248859